Basic refactoring - towards shared e2e tests
